fixed eda mV, eda 16bit and calibrate saadc

+++ eda.h
@@ -101,7 +101,7 @@ struct ble_eda_s
     uint16_t                 service_handle;            /**< Handle of Eda Service (as provided by the BLE stack). */
     ble_gatts_char_handles_t eda_level_handles;     /**< Handles related to the Eda Level characteristic. */
     uint16_t                 report_ref_handle;         /**< Handle of the Report Reference descriptor. */
-    uint8_t                  eda_level_last;        /**< Last Eda Level measurement passed to the Eda Service. */
+    uint16_t                 eda_level_last;        /**< Last Eda Level measurement passed to the Eda Service. */
     bool                     is_notification_supported; /**< TRUE if notification of Eda Level is supported. */
 };
 
@@ -147,7 +147,7 @@ void ble_eda_on_ble_evt(ble_evt_t const * p_ble_evt, void * p_context);
  * @return      NRF_SUCCESS on success, otherwise an error code.
  */
 ret_code_t ble_eda_level_update(ble_eda_t * p_eda,
-                                        uint8_t    eda_level,
+                                        uint16_t    eda_level,
                                         uint16_t    conn_handle);
